refactor: make FormDialog reusable
