+++ packages/client/src/pages/Employee/components/table/EnhancedTableToolbar.tsx
@@ -3,16 +3,16 @@
  * @param props
  */
 
-import React, { useState } from 'react';
+import { FC, useState } from 'react';
 import { alpha } from '@mui/material/styles';
-import Toolbar from '@mui/material/Toolbar';
-import Typography from '@mui/material/Typography';
-import IconButton from '@mui/material/IconButton';
-import Tooltip from '@mui/material/Tooltip';
-import DeleteIcon from '@mui/icons-material/Delete';
-import { FormDialog } from '@pages/Employee/components/form/FormDialog';
-import AddIcon from '@mui/icons-material/Add';
+import { Typography, Toolbar, IconButton, Tooltip } from '@mui/material';
+import { Delete, Add } from '@mui/icons-material';
+import { useNavigate } from 'react-router-dom';
 import { GetEmployeeListDocument, useDeleteEmployeesMutation } from '@graphql/employee/employee';
+
+import { FormDialog } from '@components/form/FormDialog';
+import { EmployeeForm } from '@pages/Employee/components/form/EmployeeForm';
+import { Paths } from '@constants/paths';
 
 interface EnhancedTableToolbarProps {
   numSelected: number;
@@ -20,15 +20,26 @@ interface EnhancedTableToolbarProps {
   setSelected: (select: readonly string[]) => void;
 }
 
-export const EnhancedTableToolbar = (props: EnhancedTableToolbarProps) => {
-  const { numSelected, selected, setSelected } = props;
+export const EnhancedTableToolbar: FC<EnhancedTableToolbarProps> = ({ numSelected, selected, setSelected }) => {
   const [open, setOpen] = useState(false);
   const [deleteEmployees, { data, loading, error }] = useDeleteEmployeesMutation();
+  const navigate = useNavigate();
 
   const handleClickOpen = () => {
     setOpen(true);
   };
 
+  /**
+   * close Model and navigate to Employee List Page
+   */
+  const handleClose = () => {
+    setOpen(false);
+    navigate(Paths.EMPLOYEE_lIST);
+  };
+
+  /**
+   * delete one or more employees
+   */
   const handleClickDelete = async () => {
     await deleteEmployees({
       variables: {
@@ -61,7 +72,7 @@ export const EnhancedTableToolbar = (props: EnhancedTableToolbarProps) => {
       {numSelected > 0 ? (
         <Tooltip title="Delete">
           <IconButton onClick={handleClickDelete}>
-            <DeleteIcon />
+            <Delete />
           </IconButton>
         </Tooltip>
       ) : (
@@ -77,9 +88,11 @@ export const EnhancedTableToolbar = (props: EnhancedTableToolbarProps) => {
               }
             }}
           >
-            <AddIcon />
+            <Add />
           </IconButton>
-          <FormDialog open={open} setOpen={setOpen} title="Add Employee" />
+          <FormDialog open={open} setOpen={setOpen} title="Add Employee" onClose={handleClose}>
+            <EmployeeForm handleClose={handleClose} />
+          </FormDialog>
         </>
       )}
     </Toolbar>
