+++ packages/client/src/pages/Project/components/table/EnhancedTableToolbar.tsx
@@ -3,16 +3,16 @@
  * @param props
  */
 
-import React, { FC, useState } from 'react';
+import { FC, useState } from 'react';
 import { alpha } from '@mui/material/styles';
-import Toolbar from '@mui/material/Toolbar';
-import Typography from '@mui/material/Typography';
-import IconButton from '@mui/material/IconButton';
-import Tooltip from '@mui/material/Tooltip';
-import DeleteIcon from '@mui/icons-material/Delete';
-import { FormDialog } from '@pages/Project/components/form/FormDialog';
-import AddIcon from '@mui/icons-material/Add';
+import { Typography, Toolbar, IconButton, Tooltip } from '@mui/material';
+import { Delete, Add } from '@mui/icons-material';
+import { useNavigate } from 'react-router-dom';
 import { useDeleteProjectsMutation, GetProjectListDocument } from '@graphql/project/project';
+
+import { FormDialog } from '@components/form/FormDialog';
+import { ProjectForm } from '@pages/Project/components/form/ProjectForm';
+import { Paths } from '@constants/paths';
 
 interface EnhancedTableToolbarProps {
   numSelected: number;
@@ -23,9 +23,18 @@ interface EnhancedTableToolbarProps {
 export const EnhancedTableToolbar: FC<EnhancedTableToolbarProps> = ({ numSelected, selected, setSelected }) => {
   const [open, setOpen] = useState(false);
   const [deleteProjects, { data, loading, error }] = useDeleteProjectsMutation();
+  const navigate = useNavigate();
 
   const handleClickOpen = () => {
     setOpen(true);
+  };
+
+  /**
+   * close Model and navigate to Track Page
+   */
+  const handleClose = () => {
+    setOpen(false);
+    navigate(Paths.PROJECT_lIST);
   };
 
   const handleClickDelete = async () => {
@@ -60,7 +69,7 @@ export const EnhancedTableToolbar: FC<EnhancedTableToolbarProps> = ({ numSelecte
       {numSelected > 0 ? (
         <Tooltip title="Delete">
           <IconButton onClick={handleClickDelete}>
-            <DeleteIcon />
+            <Delete />
           </IconButton>
         </Tooltip>
       ) : (
@@ -76,9 +85,11 @@ export const EnhancedTableToolbar: FC<EnhancedTableToolbarProps> = ({ numSelecte
               }
             }}
           >
-            <AddIcon />
+            <Add />
           </IconButton>
-          <FormDialog open={open} setOpen={setOpen} title="Add Project" />
+          <FormDialog open={open} setOpen={setOpen} onClose={handleClose} title="Add Project">
+            <ProjectForm handleClose={handleClose} />
+          </FormDialog>
         </>
       )}
     </Toolbar>
